+++ CHANGELOG.md
@@ -10,9 +10,9 @@
 ## 4.2.0 (roadmap)
 
 - [x] Deprecate `createHook` for `runActions`
-- [x] Deprecate `registerAction({ hook: 'xxx' })` for `registerAction({ action: 'xxx', name: 'yyy' })`
+- [x] Deprecate `registerAction({ hook: 'xxx' })` for `registerAction({ target: 'xxx', name: 'yyy' })`
 - [x] Deprecate `registerAction` for `registerExtension`
-- [x] Deprecate `registerHook` for `registerActions` who always get a map of hooks to add
+- [x] Deprecate `registerHook` for `registerTargets` who always get a map of hooks to add
 - [x] Deprecate `getHook()` for `getAction()`
 - [ ] Fix kitchensink/mock-axios-e2e
 - [ ] Fix service order dependency in `service-fetchq` - it should work seamlessly before/after `service-fastify`